opprydding i redaktørvarsler
