+++ packages/nextjs/src/components/_editor-only/redaktorvarsler/harRedaktorFeil.tsx
@@ -3,9 +3,9 @@ import { KortUrlVarsel } from './varsler/kort-url/KortUrlVarsel';
 import { DuplikateIder } from './varsler/duplikate-ider/DuplikateIder';
 import { SkjemanummerVarsel } from './varsler/skjemanummer/SkjemanummerVarsel';
 import { KontaktinformasjonVarsel } from './varsler/kontaktinformasjon/KontaktinformasjonVarsel';
-import { FormatertInnholdUtenforInnholdsseksjonVarsel } from './varsler/formatert-innhold-utenfor-innholdsseksjon/FormatertInnholdUtenforInnholdsseksjonVarsel';
-import { FragmentUtenforInnholdsseksjonVarsel } from './varsler/fragment-utenfor-innholdsseksjon/FragmentUtenforInnholdsseksjonVarsel';
-import { HtmlAreaInnholderDivVarsel } from './varsler/html-area-div/HtmlAreaInnholderDivVarsel';
+import { FormatertInnholdUtenforInnholdsseksjon } from './varsler/formatert-innhold-utenfor-innholdsseksjon/FormatertInnholdUtenforInnholdsseksjon';
+import { FragmentUtenforInnholdsseksjon } from './varsler/fragment-utenfor-innholdsseksjon/FragmentUtenforInnholdsseksjon';
+import { HtmlAreaInnholderDiv } from './varsler/html-area-div/HtmlAreaInnholderDiv';
 
 export const harRedaktorfeil = (content: ContentProps): boolean => {
     const isEditorView = content.editorView === 'edit' || content.editorView === 'preview';
@@ -18,8 +18,8 @@ export const harRedaktorfeil = (content: ContentProps): boolean => {
         DuplikateIder({}) !== null ||
         SkjemanummerVarsel({ content }) !== null ||
         KontaktinformasjonVarsel({ content }) !== null ||
-        FormatertInnholdUtenforInnholdsseksjonVarsel({ content }) !== null ||
-        FragmentUtenforInnholdsseksjonVarsel({ content }) !== null ||
-        HtmlAreaInnholderDivVarsel({ content }) !== null
+        FormatertInnholdUtenforInnholdsseksjon({ content }) !== null ||
+        FragmentUtenforInnholdsseksjon({ content }) !== null ||
+        HtmlAreaInnholderDiv({ content }) !== null
     );
 };
