+++ packages/nextjs/src/components/_editor-only/redaktorvarsler/varsler/formatert-innhold-utenfor-innholdsseksjon/FormatertInnholdUtenforInnholdsseksjonVarsel.tsx
@@ -1,47 +1,10 @@
-import React from 'react';
-import { ContentProps } from 'types/content-props/_content-common';
-import { pageContentFormatertInnholdUtenforInnholdsseksjon } from './pageContentFormatertInnholdUtenforInnholdsseksjon';
-
 type Props = {
-    content: ContentProps;
+    warnings: React.ReactNode;
     className?: string;
 };
 
-export const FormatertInnholdUtenforInnholdsseksjonVarsel = ({ content, className }: Props) => {
-    const warnings: React.ReactElement[] = [];
-
-    const finnHtmlAreaUtenforInnholdsseksjon = (node: any): void => {
-        if (!node || typeof node !== 'object') return;
-
-        if (pageContentFormatertInnholdUtenforInnholdsseksjon(node)) {
-            const { path, config } = node;
-
-            if (!config.html?.processedHtml) {
-                return;
-            }
-
-            warnings.push(
-                <ul key={`${path}-list`}>
-                    <li key={`${path}-item`}>
-                        Innhold: {JSON.stringify(config.html.processedHtml)}
-                    </li>
-                </ul>
-            );
-        }
-
-        if (Array.isArray(node)) {
-            node.forEach(finnHtmlAreaUtenforInnholdsseksjon);
-        } else {
-            Object.values(node).forEach(finnHtmlAreaUtenforInnholdsseksjon);
-        }
-    };
-
-    const regions = content.page?.regions;
-    if (regions && 'pageContent' in regions) {
-        finnHtmlAreaUtenforInnholdsseksjon(regions['pageContent']);
-    }
-
-    return warnings.length > 0 ? (
+export const FormatertInnholdUtenforInnholdsseksjonVarsel = ({ className, warnings }: Props) => {
+    return (
         <li key="part-utenfor-innholdsseksjon-warning" className={className}>
             Innholdet ligger utenfor den angitte innholdsseksjonen, noe som kan føre til
             visningsfeil på nav.no. Under ser du hvilket innhold det gjelder.
@@ -54,5 +17,5 @@ export const FormatertInnholdUtenforInnholdsseksjonVarsel = ({ content, classNam
                 </li>
             </ul>
         </li>
-    ) : null;
+    );
 };
