+++ packages/nextjs/src/components/_editor-only/redaktorvarsler/varsler/html-area-div/HtmlAreaInnholderDiv.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { ContentProps } from 'types/content-props/_content-common';
 import { htmlAreaInneholderDiv } from './htmlAreaInneholderDiv';
+import { HtmlAreaInneholderDivVarsel } from './HtmlAreaInneholderDivVarsel';
 
 type Props = {
     content: ContentProps;
     className?: string;
 };
 
-export const HtmlAreaInnholderDivVarsel = ({ content, className }: Props) => {
+export const HtmlAreaInnholderDiv = ({ content, className }: Props) => {
     const warnings: React.ReactElement[] = [];
 
     const finnDivTagsIFormatertInnhold = (node: any): void => {
@@ -38,21 +39,6 @@ export const HtmlAreaInnholderDivVarsel = ({ content, className }: Props) => {
     }
 
     return warnings.length > 0 ? (
-        <li key="html-area-div-warning" className={className}>
-            Det er brukt en <code>&lt;div&gt;</code>-tagg i seksjonen for formatert innhold, noe som
-            kan føre til visningsfeil på nav.no. Under ser du hvilket innhold det gjelder.
-            {warnings}
-            <strong>Slik retter du feilen:</strong>
-            <ul>
-                <li key="rette-feilen-1">
-                    Lim innholdet i en ren tekst-editor (f.eks. Notepad). Kopier teksten derfra og
-                    lim den inn i Enonic. Dette fjerner skjult formatering og uønsket kode.
-                </li>
-                <li>
-                    Hvis du kan HTML: Fjern <code>&lt;div&gt;</code>-taggen direkte i
-                    kildekodefeltet i Enonic.
-                </li>
-            </ul>
-        </li>
+        <HtmlAreaInneholderDivVarsel className={className} warnings={warnings} />
     ) : null;
 };
